Add cabal support
Problem:
I've met problem with one package version. I use cabal to build project and cabal for some reason decided to use old version of this package. With stack everything fine.
So, I think not only me, but other people prefer cabal over stack, so we should add official cabal support.

Solution:
Store .cabal and generate cabal.project and cabal.project.freeze with stack2cabal. To be sure that we do not forgot to run this command we have to add check to our ci.